Create randomlyAddTaxa
This is the first commit of this file. I do not have direct access to the repo to add it myself. Note that the #' symbols at the start of the file are roxygen comments, and will automatically generated an .Rd help file if we build any packages in that way.