Major change to functionality of randomlyAddTaxa
Two major changes to function, plus a minor change. Minor change: added better example. Major change 1: tree is now built up iteratively, so that new species can be bound to either a species in the original phylogeny, or to a species that was added in subsequently. Major change 2: can now choose to add species either above, below, in a polytomy or randomly above or below the node in question.