Fixed bug: adding taxa stem-wards
There is a check built into this function that if a single taxon is sister to all others, and if "stem" or "randomly" is chosen, any taxa to be added will always be added crown-wards to avoid function failure by trying to add the taxa below the root. Fixed incorrect calculation of whether the first part of this if statement was true.